refactor(test): split hook_test.go by source file and merge redigo helper tests
Per review feedback: avoid standalone test files detached from the
source file they cover.

- goredis: split hook_test.go into redis_error_test.go (helper unit
  tests for redisSpanEndErr / redisPipelineSpanEndErr / getRedisV9Statement)
  and setup_test.go (ProcessHook / ProcessPipelineHook span-status tests)
- goredisv8: same split into redis_error_test.go and setup_test.go
- redigo: merge TestRedigoSpanEndErr from redis_error_test.go into
  conn_test.go; delete the now-empty redis_error_test.go

rueidisnil/nil_test.go is kept as-is: the rueidisnil sub-package exists
specifically because pkg/rules/rueidisgo cannot be unit-tested directly
(it references rueidis.RedisAdders only available via bytecode injection),
so nil_test.go is the only viable location for these helper tests.
